requirements.txt
ansicolors==1.1.8
appdirs==1.4.3
argon2-cffi==19.1.0
asgiref==3.2.5
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
cryptography==2.9.2
distlib==0.3.0
Django==3.0.3
django-address==0.2.2
django-filter==2.2.0
django-gulp==4.1.0
django-ipware==2.1.0
django-netfields==1.2.2
django-phone-field==1.8.0
django-taggit==1.2.0
env-tools==2.2.0
Faker==4.0.3
filelock==3.0.12
geoip2==3.0.0
idna==2.9
importlib-metadata==1.6.0
lxml==4.5.0
maxminddb==1.5.4
more-itertools==8.2.0
netaddr==0.7.19
olefile==0.46
packaging==20.3
Pillow==7.1.0
pluggy==0.13.1
psutil==5.7.0
py==1.8.1
pycparser==2.20
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytracking==0.2.0
pytz==2019.3
requests==2.23.0
six==1.11.0
sqlparse==0.3.0
text-unidecode==1.3
tini==3.0.1
toml==0.10.0
tox==3.15.0
urllib3==1.25.9
virtualenv==20.0.20
wcwidth==0.1.9
wincertstore==0.2
yagmail==0.11.224
zipp==3.1.0
